🚚 [rename] New 템플릿 이름변경 #86
