+++ src/molecules/MyPostTemplate/MyPostTemplate.jsx
@@ -1,44 +1,46 @@
-import Masonry from 'react-masonry-css';
-import { motion } from 'framer-motion';
-import { useNavigate, useLocation } from 'react-router-dom';
-import { useImageStore } from '@/zustand/useStore';
+import Masonry from 'react-masonry-component';
+import useResizeUpdateView from '@/hooks/useResizeUpdateView';
 import TextContents from '@/atoms/TextContents/TextContents';
+import { motion } from 'framer-motion';
+import { useLocation, useNavigate, useParams } from 'react-router-dom';
+import { useImageStore } from '@/zustand/useStore';
 
-function MyImageTemplate({ images }) {
+function MyPostTemplate({ images }) {
   const navigate = useNavigate();
   const location = useLocation();
   const setSelectedImageUrl = useImageStore((state) => state.setSelectedImageUrl);
+  const { boardText } = useParams();
 
-  const onBoxClicked = (imageId, imageUrl, category) => {
-    // if (location.pathname.startsWith('/mypage/post')) {
-    //   navigate(`/mypage/post/${category}/detail/${imageId}`);
-    // }
+  // 창 크기 조정에 따라 컴포넌트 리-렌더링
+  useResizeUpdateView();
+
+  const onBoxClicked = (imageId, imageUrl) => {
     if (location.pathname.startsWith('/mypage/post')) {
-      navigate(`/community/detail`);
+      navigate(`/mypage/post/detail/${imageId}`, { state: { imageSrc: imageUrl } });
+    } else if (location.pathname.includes('all')) {
+      navigate(`/mypage/bookmark/field/all/detail/${imageId}`, { state: { imageSrc: imageUrl } });
+    } else if (boardText) {
+      navigate(`/mypage/bookmark/${boardText}/detail/${imageId}`, { state: { imageSrc: imageUrl } });
     }
     setSelectedImageUrl(imageUrl);
   };
 
-  const breakpointColumnsObj = {
-    default: 4,
-    639: 3,
-    450: 2,
-  };
-
   return (
-    <ul className={`max-w-screen-md h-auto min-h-[600px] bg-white`}>
-      <Masonry
-        breakpointCols={breakpointColumnsObj}
-        className="my-masonry-grid flex gap-[12px]"
-        columnClassName="my-masonry-grid_column min-w-[170px] flex flex-col items-center"
-      >
-        {images.map(({ id, imageUrl, title, category }, index) => (
-          <motion.li key={id} layoutId={id} className="h-[350px]">
-            <img
-              src={imageUrl}
-              className="w-[170px] h-[270px] bg-gray-100 rounded-2xl cursor-zoom-in object-cover"
-              onClick={() => onBoxClicked(id, imageUrl, category)}
-            />
+    <Masonry
+      elementType="ul"
+      className={`min-w-[320px] min-h-[600px] max-w-screen-md h-auto bg-white relative`}
+      options={{
+        gutter: 15,
+        horizontalOrder: true,
+        itemSelector: '.masonry-item',
+        transitionDuration: 0,
+        fitWidth: true,
+      }}
+    >
+      {images.map(({ id, imageUrl, title }, index) => (
+        <motion.li key={id} layoutId={id} className="masonry-item absolute h-[350px]" style={{ margin: '-2px' }}>
+          <button type="button" onClick={() => onBoxClicked(id, imageUrl)}>
+            <img src={imageUrl} className={`w-[170px] h-[260px] bg-gray-100 rounded-2xl cursor-zoom-in object-cover`} />
             <TextContents
               text={title}
               maxLength={25}
@@ -46,16 +48,16 @@ function MyImageTemplate({ images }) {
               fontWeight="font-bold"
               fontSize="text-[18px]"
               padding="p-2"
-              margin="mb-4"
               style={{
                 width: '170px',
                 textAlign: 'center',
               }}
             />
-          </motion.li>
-        ))}
-      </Masonry>
-    </ul>
+          </button>
+        </motion.li>
+      ))}
+    </Masonry>
   );
 }
-export default MyImageTemplate;
+
+export default MyPostTemplate;
